Deleted metrics.py and restored code in util.py
